+++ src/test/groovy/com/commercehub/gradle/plugin/avro/BuildCacheSupportFunctionalSpec.groovy
@@ -80,12 +80,12 @@ class BuildCacheSupportFunctionalSpec extends FunctionalSpec {
         given: "a project is built once with build cache enabled"
         copyResource("mail.avpr", avroDir)
         buildFile << """
-            task("generateSchema", type: com.commercehub.gradle.plugin.avro.GenerateAvroSchemaTask) {
-                source file("src/main/avro")
-                include("**/*.avpr")
-                outputDir = file("build/generated-main-avro-avsc")
-            }
-        """
+        |tasks.register("generateSchema", com.commercehub.gradle.plugin.avro.GenerateAvroSchemaTask) {
+        |    source file("src/main/avro")
+        |    include("**/*.avpr")
+        |    outputDir = file("build/generated-main-avro-avsc")
+        |}
+        |""".stripMargin()
         run("generateSchema", "--build-cache")
 
         and: "the project is cleaned"