+++ CHANGES.md
@@ -9,6 +9,7 @@
 * Add support for testing multiple Kotlin versions
 * Update plugin's own build to address some deprecation warnings of APIs being removed in Gradle 7
 * Add tests for Kotlin DSL usage (#61)
+* Support [Task Configuration Avoidance](https://docs.gradle.org/current/userguide/task_configuration_avoidance.html) (#97); thanks to [dcabasson](https://github.com/dcabasson) for the collaboration
 
 ## 0.18.0
 * Use reproducible file order for plugin archives